remove try/catch in adding project modal because we now enable duplicate names (suffix get added). add module in helpers/ folder to check for duplicate project names and add suffix if required
