+++ src/models/Project.js
@@ -1,3 +1,4 @@
+import addSuffixToDuplicateProjectName from "../helpers/addSuffixToDuplicateProjectNames";
 import {
   getCleanedProjectNames,
   retrieveProjectNames,
@@ -22,25 +23,9 @@ const createProject = (recreatingFromJSON = false) => {
 
   // SETTERS
   const setName = (retrievedOrNewName) => {
-    const projectNamesInStorage = retrieveProjectNames();
-    const cleanedProjectNamesInStorage = getCleanedProjectNames();
-    const projectNameExists =
-      projectNamesInStorage.includes(retrievedOrNewName);
-
-    if (!projectNameExists || recreatingFromJSON)
-      return (name = retrievedOrNewName);
-
-    console.log("project name exists");
-
-    const duplicateCount = cleanedProjectNamesInStorage.filter(
-      (n) => n === retrievedOrNewName
-    ).length;
-
-    for (let suffixInt = 1; suffixInt <= duplicateCount; suffixInt++) {
-      const projectNameWithSuffix = `${retrievedOrNewName} (${suffixInt})`;
-      if (!projectNamesInStorage.includes(projectNameWithSuffix))
-        return (name = projectNameWithSuffix);
-    }
+    if (recreatingFromJSON) return (name = retrievedOrNewName);
+    name = addSuffixToDuplicateProjectName(retrievedOrNewName);
+    console.log(name);
   };
 
   const setId = (retrievedId) => {
